Fixed some minor security issue.

+++ src/main/java/com/googlecode/paradox/metadata/tables/ViewColumnUsage.java
@@ -92,7 +92,7 @@ public class ViewColumnUsage implements Table {
             int sum = 0;
 
             for (final Schema localSchema : connectionInfo.getSchemas(catalogName, null)) {
-                sum += localSchema.list(connectionInfo, null).stream()
+                sum += (int) localSchema.list(connectionInfo, null).stream()
                         .filter(table -> (table instanceof View)).count();
             }
